Création d'un article manager qui contient les fonctions liés à article + hydratation de la classe Article. Précédemment les 2 étaient dans un seul et unique fichier.
